requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
et-xmlfile==1.1.0
idna==3.8
msal==1.31.0
numpy==2.1.1
Office365-REST-Python-Client==2.5.12
openpyxl==3.1.5
pandas==2.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
termcolor==2.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
XlsxWriter==3.2.0
